refactor(collector): split the three phases the pass always had
`Collector` 833 -> 393. The pass runs in three phases and always did — the
lexical walk fills the shapes, the corpus walk merges other files' into them,
the resolution reads them — but only the first had a class. The boundary
between the last two was already written down, as a `protected attr_reader`
listing fourteen collections and an `absorb` concatenating them one by one.
This gives both a name.

`ShapeSet` (53 lines)

  The fourteen collections that cross the boundary. `absorb`'s fifteen lines of
  hand-written `concat` become `@shapes.merge(other.shapes)`, and the fourteen
  protected readers become one.

  `apply_calls` is the one asymmetric merge, and the asymmetry is the point:
  another file's call sites are read for one question — which modules it
  registered on a waypoint — and never emitted from, so they land in
  `foreign_applies` where nothing that emits will look.

`Resolution` (442 lines)

  Seventeen methods and two tables: a `ShapeSet` and a `Declarations` in,
  replays and extensions out, and no way to reach back into the walk that
  produced them.

`Collector#resolve_shapes` is the seam between phase two and three, and naming
it fixed an ordering this pass had been getting away with. `DslGraph` was built
inside `resolve_replays` from `@resolved_delegations` and the delegations were
concatenated onto that array AFTERWARDS — correct only because it was the same
object, mutated in place. The resolution now starts after collecting finishes,
so nothing depends on a write landing after a read. Both comments that claimed
otherwise are updated rather than left to mislead.

`initialize` goes from 26 instance variables to 10, and what is left is the
collection's own: the raw shapes waiting to be resolved against the file's
declarations.

Suite: 1546 examples, 3 failures — the three worktree-environment ones that
also fail on main from a worktree and pass in a normal checkout.
